refactor(services): remove Service suffix from services files
