+++ server/src/services/achievements/index.ts
@@ -1,4 +1,4 @@
-export * from "./achievementsService";
+export * from "./achievements";
 export * from "./achivementUserProgresses";
 export * from "./achievementStages";
 export * from "./types";
