+++ server/src/services/affixes/index.ts
@@ -1,3 +1,3 @@
-export * from "./affixesService";
+export * from "./affixes";
 export * from "./";
 export type { AffixCreateData, AffixUpdateData } from "./types";
