+++ server/src/services/auth/index.ts
@@ -1,4 +1,4 @@
-export * from "./authService";
+export * from "./auth";
 export * from "./";
 
 export type { AuthCreateData } from "./types";
