+++ server/src/services/chatCommands/index.ts
@@ -1,4 +1,4 @@
-export * from "./chatCommandsService";
+export * from "./chatCommands";
 export * from "./";
 
 export type { ChatCommandCreateData, ChatCommandUpdateData } from "./types";
