+++ server/src/services/configs/index.ts
@@ -1,4 +1,4 @@
-export * from "./configsService";
+export * from "./configs";
 export * from "./";
 
 export type { ConfigUpdateData } from "./types";
