+++ server/src/services/messageCategories/index.ts
@@ -1,4 +1,4 @@
-export * from "./messageCategoriesService";
+export * from "./messageCategories";
 export * from ".";
 
 export type { MessageCategoryCreateData } from "./types";
