+++ server/src/services/moods/index.ts
@@ -1,4 +1,4 @@
-export * from "./moodsService";
+export * from "./moods";
 export * from "./";
 
 export type { MoodCreateData, MoodUpdateData } from "./types";
