+++ server/src/services/overlays/index.ts
@@ -1,4 +1,4 @@
-export * from "./overlaysService";
+export * from "./overlays";
 export * from "./";
 
 export { OverlayCreateData, OverlayUpdateData } from "./types";
