+++ server/src/services/redemptions/index.ts
@@ -1,4 +1,4 @@
-export * from "./redemptionsService";
+export * from "./redemptions";
 export * from "./";
 
 export { RedemptionCreateData } from "./types";
