+++ server/src/services/songs/index.ts
@@ -1,4 +1,4 @@
-export * from "./songsService";
+export * from "./songs";
 export * from ".";
 
 export { SongsCreateData, SongsUpdateData, ManageSongLikesAction } from "./types";
