+++ server/src/services/streamSessions/index.ts
@@ -1,2 +1,2 @@
-export * from "./streamSessionsService";
+export * from "./streamSessions";
 export * from "./";
