+++ server/src/services/tags/index.ts
@@ -1,4 +1,4 @@
-export * from "./tagsService";
+export * from "./tags";
 export * from "./";
 
 export { TagCreateData, TagUpdateData } from "./types";
