+++ server/src/services/timers/index.ts
@@ -1,4 +1,4 @@
-export * from "./timersService";
+export * from "./timers";
 export * from "./";
 
 export { TimerCreateData, TimerUpdateData } from "./types";
