+++ server/src/services/triggers/index.ts
@@ -1,4 +1,4 @@
-export * from "./triggersService";
+export * from "./triggers";
 export * from "./";
 
 export { TriggerCreateData, TriggerUpdateData } from "./types";
