+++ server/src/services/widgets/index.ts
@@ -1,4 +1,4 @@
-export * from "./widgetsService";
+export * from "./widgets";
 export * from "./";
 
 export { WidgetCreateData, WidgetUpdateData } from "./types";
